Make python narupa a "namespace package"
A namespace package is a single namespace that can be populated by
multiple packages. This means that we can have a "narupa-core" package
that will provide "narupa.core", and an other package (e.g.
narupa-openmm) that will provide narupa.openmm.

Being able to split the package this way allows to have more independant
development while keeping the same namespace and branding.