Move python library from narupy to narupa
The python library and services were split between two namespaces:
narupa, and narupy. While the name "narupy" is brilliant, it is
sub-optimal from a user and API point of view to have these two namespaces.

This commit moves the functional parts of narupy in the narupa python
module. As moving files around tend to make diffs impossible to review,
this commit mostly focus on moving files, and included only very limited
changes to the code. These changes are all fixes in the imports to
reflect the new naming.

This commit is a first step toward the organisation suggested in
<https://gitlab.com/intangiblerealities/narupa-protocol/issues/1#note_152601533>.