Fix memory leak and add locks
The memory leak described in #35 was due to client queues not being
cleared after a client disconnect. Without client to consume the frames
in that queue, the queue grows without control. This commit makes sure
the queues are cleared when clients disconnect.

While debugging, I encounter a couple of race conditions so I also added
locks in the frame publisher.