Refactor the test for the timeout in ESSD search
The test was relying on pytest.mark.timeout, though timeout failure are
catastrophic.

Here, we refactor the test so we actually test for the duration of the
search. If the test is too long, then the test is considered failing.
The timeout mark is kept to avoid tests hanging in the worst case
scenario.

The search duration is also extended for that specific test. A search
time shorter than a second makes it likely to encounter an edge case
that case the search time to be about twice the required one. While the
ideal solution would be to fix the edge case, it is probably not worth
fixing.


Former-commit-id: 955a597a6d34e886b13a9d0d23af536535d4c2a6