Fix the ESSD search time
In the previous commit, I extended the search time when testing the
timeout of the essd search. This was due to a percieived edge case on
short timeouts. This was actually a very fixable error in the logic
where we were waiting for the whole search time at every search
iteration, causing the actual time in the function to be doubled if
nothing was found during the first iteration. It appeared as an edge
case because the likelyhood of finding a server during the first search
iteration decreases as the search time reduces.


Former-commit-id: de7a15833db14456271b0d307b86da23ed80fce4