NarupaImdClient doesn't subscribe to frames by default
A client created using `narupa.app.NarupaImd` does not subscribe to
frames by default anymore. This saves bandwidth for usages that do not
require access to the frames.

To subscribe to the frames, the client needs to call the
`subscribe_to_frames` or `subscribe_to_all_frames` methods. The former
let's the user set a sending interval, the latter does subscribe to all
the frames. Until one of these methods is called, the methods and
attributes that access the frames raise a RuntimeError with an explicit
message.

The `all_frames` argument of the client's `__init__` is deprecated and a
warning is issued if it is set. The attribute has no effect anymore,
this should not be an issue as it was probably never used.

The `all_frames` attribute now returns `False` if the client does not
subscribe to frames or if the client subscribed using
`subscribe_to_frames`, and `True` if the client subscribed using
`subscribe_to_all_frames`.


Former-commit-id: 19c74bdd01ac6b2b8697a7e3da0cd1383a21a7d7