Do not reset the user forces when resetting
I made the reset of the openmm runner zero out the user interactions
forces as I initially thought it was causing the system to explode just
after a reset. However, this was not the cause and zeroing these forces
means that a sustained user interaction happening during the reset will
be lost. This would happen if a user interaction is created with the
python client.


Former-commit-id: ffe9851d8849e530efe298609afa7f5bef32c87b