Run the tests on github
The python and dotnet tests run as github CI workflows. This means there
is a minimum amount of CI going on the github repository.

Note: not all the gitlab CI pipeline is ported yet! The conda packages
are neither built, tested, not published. The minimum version of the
dependancies is not tested. The python types are not validated.