Allow to choose the OpenMM platform in narupa-ase
By default, OpenMM chooses the most performant platform available.
However, if there is an issue with that platform, OpenMM will fail
leaving narupa broken without ways to recover.

This commit adds the option to ask for a specific platform when
deserialising an xml simulation and adds an option to the narupa-omm-ase
command line.