OpenMM runner fails with more verbosity
When the simulation fails in the OpenMM runner, it may appear as just a
freeze with no explanation. This is because the simulation would fail on
a separate thread. This commit adds prints in the simulation thread, so
the cause of the failure appears somewhere.

This is a workaround and it should be solved in a better way. If the
simulation fails, it should cause the server to stop with a proper error
message.