Read a trajectory recording as an MDAnalysis Universe
MDAnalysis is a powerful tool to analyse MD trajectories. However, there
are only limited facilities to allow loading a Narupa trajectory into an
MDAnalysis Universe. Indeed, `narupa.mdanalysis.converter` can create a
Universe from a FrameData, but it cannot work with a trajectory,
especially not a trajectory as recorded by the rust server or by
nanover-recorder.

This commit adds the `narupa.mdanalysis.universe` module that implements
a reader and a parser for Narupa trajectory recordings.