Fix finding new topologies in narupa.mdanalysis.universe
MDAnalysis universes cannot see their topology change so we need to cut
the trajectory where the topology is modified. We also issue a warning
if we do have to exclude frames. The test to issue the warning was
incorectly placed, consuming the first frame of the trajectory instead
of consuming the first frame *after* we stop reading the trajectory.

This commit moves the test to get the expected behaviour.