Refactor mda trajectory notebook
I checked that this notebook does what it is supposed to do, including testing viewing and running the play/pause/reset commands in VR (all working fine).

This commit includes some minor refactoring, a bug fix for an outdated matplotlib function, and addition of some documentation/comments.

Renamed the notebook to make it's function clearer, but maybe this isn't useful?