Update README (#234)
* Update Tutorials section

Replace the long text about the different tutorials as this is outdated. Reduce the text by replacing with a link to the examples folder in the repo and the tutorials page of the documentation.

* Add link to docs for running a server

Add a link to the docs page about how to run a NanoVer server.

* Split installation instructions

Split more clearly into sections for user installation & developer installation. In the developer installation, split the sections more clearly into Mac/Linux & Windows (putting windows first as this is probs the most common OS for our users).

* Separate completely dev & user installation guides

Put the user and developer installation as the main headings, reformatting the rest of the headings with this change.

* Fix typos

* Update intro

Remove old "2", leftover from fork off Narupa. Remove duplicate mention of NanoverImd.

* Add table of contents

Add table of contents with main headings for easier navigation.

* Update wording of intro

* Update README.md

Add links to information that would be useful for developers running the NanoVer examples in the developer guide.

* Update heading

Remove "2"

* Minor reformatting

One of the subheadings was at the wrong level. Added some intro to the `getting started` section. Fixed typo.

* Update subheading levels

Move Running the tests & Running the examples to their own subheadings, rather than being nested in the Developer instructions. Add these new headings to the contents.

* Change conda install to conda update