Update the notebooks based on the repo name changes
Update the jupyter notebooks to reflect the name changes of the nanover imd repo. In particular, I changed instances of "NanoVer iMD" & "NanoVerIMD" to "NanoVer iMD-VR".

I also replaced an (old) url to the nanover-imd repo in mdanalysis_trajectory notebook with a link to the docs. This keeps it consistent with the other notebooks.